# lib/api/openapi.yaml
openapi: 3.0.3
info:
  title: RestaurantOS API
  description: |
    Full REST API for RestaurantOS v2 Restaurant Management System.
    All protected endpoints require a Bearer JWT token.
  version: 2.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /tables:
    get:
      tags: [Tables]
      security: [{ BearerAuth: [] }]
      summary: List all tables
      responses:
        '200': { description: Tables list }
  /tables/{id}/status:
    put:
      tags: [Tables]
      security: [{ BearerAuth: [] }]
      summary: Update table status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Status updated }
...
  /reservations:
    get:
      tags: [Reservations]
      security: [{ BearerAuth: [] }]
      summary: List reservations
      parameters:
        - name: date
          in: query
          schema: { type: string, format: date }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { description: Reservations list }
    post:
      tags: [Reservations]
      security: [{ BearerAuth: [] }]
      summary: Create reservation
      responses:
        '201': { description: Reservation created }
  /reservations/{id}/status:
    put:
      tags: [Reservations]
      security: [{ BearerAuth: [] }]
      summary: Update reservation status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Status updated }
...
  /staff:
    get:
      tags: [Staff]
      security: [{ BearerAuth: [] }]
      summary: Staff list (Admin/Manager)
      responses:
        '200': { description: Staff list }
  /staff/{id}/status:
    put:
      tags: [Staff]
      security: [{ BearerAuth: [] }]
      summary: Update staff status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Status updated }
  /staff/{id}/deactivate:
    put:
      tags: [Staff]
      security: [{ BearerAuth: [] }]
      summary: Deactivate account (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Account deactivated }
...